Create dir for each platform
