+++ forge/build.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    id("com.kotori316.common.java")
+    id("com.kotori316.common.publish")
+    id("com.kotori316.common.signing")
+}
