+++ gradle.properties
@@ -4,9 +4,7 @@ org.gradle.jvmargs=-Xmx4G
 org.gradle.daemon=true
 org.gradle.parallel=true
 org.gradle.caching=true
-
-tu_archives_base_name=test_utility
-du_archives_base_name=debug_utility
+archives_base_name=debug_utility
 minecraft=1.21.1
 version=21.1.2
 maven_latest=21.1.99-SNAPSHOT
